Fix circular carousels with scoll != 1
When scroll was not 1, rotation did odd things around the ends. this is now
fixed. As part of the fix, this.itemLength now is consitently the number of
items given not including any circular buffers. code is a lot more verbose but
but a lot simpler now as well.

Other new features include _set method for directly setting the location and this.slide alias for ul. These fix some other smaller bugs and simplify a lot of logic.