Fix bug where set method called to early.
The set method called durring the rebuild process is called before a lot
of information that would be used to actually set the placement is
available. We'll just rely on refresh even though there is a bit of a
chance for things to be visible very shortly at the wrong position.